test(plugin-page-builder): move the blocks validator suite to the plugin
